feat: update balances on deposit
